Add retry logic to RetireActualLRPs
RemoveActualLRP (which is called by RetireActualLRPs for
UNCLAIMED/CRASHED actual LRPs has a race condition:

It will get the actual LRP from the BBS and try to compare and delete it,
but it's possible that something else has already compared and swapped
the LRP in between those two calls, making the CAD fail.

This retry logic is meant to retry retiring the actual LRP in the event
of receiving ErrStoreComparisonFailed.

[#87860972]

Signed-off-by: Ashvin Agrawal <ashvin@pivotal.io>